Moved interaction and ready inside functions
