+++ src/functions/ready.ts
@@ -1,6 +1,6 @@
-import { token, guildId, clientId } from './bot_config';
+import { token, guildId, clientId } from '../bot_config';
 import { Client, REST, Routes } from 'discord.js';
-import { commands } from './commands/commands';
+import { commands } from '../commands/commands';
 
 export async function onReady(client: Client) {
   const rest = new REST({ version: '10' }).setToken(token);
