+++ src/main.ts
@@ -1,8 +1,9 @@
 import { Client, Collection, Events, GatewayIntentBits } from 'discord.js';
+import { connect, set } from 'mongoose';
 import { token } from './bot_config';
 import { commands } from './commands/commands';
-import { onInteraction } from './interaction';
-import { onReady } from './ready';
+import { onInteraction } from './functions/interaction';
+import { onReady } from './functions/ready';
 
 const client = new Client({
   intents: [GatewayIntentBits.Guilds],
@@ -26,6 +27,15 @@ client.once(Events.ClientReady, async (client) => {
 
 client.on(Events.InteractionCreate, async (interaction) => {
   await onInteraction(interaction);
+});
+
+set('strictQuery', true);
+
+connect('mongodb://localhost:27017/vynos', {}, (err) => {
+  if (err)
+    console.log(err);
+  else
+    console.log('mongdb is connected');
 });
 
 client.login(token);
